LPS-66935 Move asset categories admin related data handlers to the proper packages
